Updated module augmentation for alignment package.
