+++ packages/ckeditor5-alignment/src/alignmentcommand.ts
@@ -109,9 +109,3 @@ function setAlignmentOnSelection( blocks: Array<Element>, writer: Writer, alignm
 		writer.setAttribute( ALIGNMENT, alignment, block );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface CommandsMap {
-		alignment: AlignmentCommand;
-	}
-}
