+++ packages/ckeditor5-alignment/src/alignmentconfig.ts
@@ -74,15 +74,3 @@ export type AlignmentFormat = {
 };
 
 export type SupportedOption = 'left' | 'right' | 'center' | 'justify';
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface EditorConfig {
-
-		/**
-		 * The configuration of the {@link module:alignment/alignment~Alignment alignment feature}.
-		 *
-		 * Read more in {@link module:alignment/alignmentconfig~AlignmentConfig}.
-		 */
-		alignment?: AlignmentConfig;
-	}
-}
